Updated planning-tools nearby airport routes to allow passing an IP or a location

+++ planning-tools/src/main/kotlin/com/pilothelper/service/AirportService.kt
@@ -105,7 +105,7 @@ class AirportService(
             .map { it.toAirport() }
     }
 
-    suspend fun findAllNearby(point: Coordinates, radius: Int): List<Airport> = dbQuery {
+    suspend fun findAllNearby(point: Coordinates, radius: Float): List<Airport> = dbQuery {
         Airports.selectAll().where {
             Airports.latitude.between(
                 point.latitude - radius,
